+++ back/test.sh
@@ -362,25 +362,27 @@ ipv6_tunnel() {
         update_sysctl "net.ipv6.conf.server-ipv6-${ipv6_subnets_used_num}.proxy_ndp=1"
         update_sysctl "net.ipv6.conf.all.accept_ra=2"
         $sysctl_path -p
-        if [ "$system_arch" = "x86" ]; then
-            wget ${cdn_success_url}https://github.com/spiritLHLS/pve/releases/download/ndpresponder_x86/ndpresponder -O /usr/local/bin/ndpresponder
-            wget ${cdn_success_url}https://raw.githubusercontent.com/spiritLHLS/pve/main/extra_scripts/ndpresponder.service -O /etc/systemd/system/ndpresponder.service
-            chmod 777 /usr/local/bin/ndpresponder
-            chmod 777 /etc/systemd/system/ndpresponder.service
-        elif [ "$system_arch" = "arch" ]; then
-            wget ${cdn_success_url}https://github.com/spiritLHLS/pve/releases/download/ndpresponder_aarch64/ndpresponder -O /usr/local/bin/ndpresponder
-            wget ${cdn_success_url}https://raw.githubusercontent.com/spiritLHLS/pve/main/extra_scripts/ndpresponder.service -O /etc/systemd/system/ndpresponder.service
-            chmod 777 /usr/local/bin/ndpresponder
-            chmod 777 /etc/systemd/system/ndpresponder.service
-        fi
-        if [ -f "/usr/local/bin/ndpresponder" ]; then
-            new_exec_start="ExecStart=/usr/local/bin/ndpresponder -i ${interface} -n ${ipv6_address_without_last_segment}/${ipv6_prefixlen}"
-            file_path="/etc/systemd/system/ndpresponder.service"
-            line_number=6
-            sed -i "${line_number}s|.*|${new_exec_start}|" "$file_path"
-            systemctl start ndpresponder
-            systemctl enable ndpresponder
-            systemctl status ndpresponder 2>/dev/null
+        if [ ! -f /usr/local/bin/ndpresponder ] && [ ! -f /etc/systemd/system/ndpresponder.service ]; then
+            if [ "$system_arch" = "x86" ]; then
+                wget ${cdn_success_url}https://github.com/spiritLHLS/pve/releases/download/ndpresponder_x86/ndpresponder -O /usr/local/bin/ndpresponder
+                wget ${cdn_success_url}https://raw.githubusercontent.com/spiritLHLS/pve/main/extra_scripts/ndpresponder.service -O /etc/systemd/system/ndpresponder.service
+                chmod 777 /usr/local/bin/ndpresponder
+                chmod 777 /etc/systemd/system/ndpresponder.service
+            elif [ "$system_arch" = "arch" ]; then
+                wget ${cdn_success_url}https://github.com/spiritLHLS/pve/releases/download/ndpresponder_aarch64/ndpresponder -O /usr/local/bin/ndpresponder
+                wget ${cdn_success_url}https://raw.githubusercontent.com/spiritLHLS/pve/main/extra_scripts/ndpresponder.service -O /etc/systemd/system/ndpresponder.service
+                chmod 777 /usr/local/bin/ndpresponder
+                chmod 777 /etc/systemd/system/ndpresponder.service
+            fi
+            if [ -f "/usr/local/bin/ndpresponder" ]; then
+                new_exec_start="ExecStart=/usr/local/bin/ndpresponder -i ${interface} -n ${ipv6_address_without_last_segment}/${ipv6_prefixlen}"
+                file_path="/etc/systemd/system/ndpresponder.service"
+                line_number=6
+                sed -i "${line_number}s|.*|${new_exec_start}|" "$file_path"
+                systemctl start ndpresponder
+                systemctl enable ndpresponder
+                systemctl status ndpresponder 2>/dev/null
+            fi
         fi
         _yellow "This tunnel will use ${tunnel_mode} type"
         _yellow "这个通道将使用${tunnel_mode}类型"
@@ -431,12 +433,19 @@ else
     _yellow "${tunnel_mode} 协议不符合规则"
     exit 1
 fi
-
 if [ ! -d /usr/local/bin ]; then
     mkdir -p /usr/local/bin
 fi
 statistics_of_run-times
 _green "脚本当天运行次数:${TODAY}，累计运行次数:${TOTAL}"
+if [ -f /usr/local/bin/6in4_usable_subnets ]; then
+    ipv6_subnets_usable_num=$(cat /usr/local/bin/6in4_usable_subnets | grep "^Network" | wc -l)
+    _blue "The number of ${target_mask} subnets available: ${ipv6_subnets_usable_num}"
+fi
+if [ -f /usr/local/bin/6in4_used_subnets ]; then
+    ipv6_subnets_used_num=$(cat /usr/local/bin/6in4_used_subnets | grep "^Network" | wc -l)
+    _blue "The number of ${target_mask} subnets used: ${ipv6_subnets_used_num}"
+fi
 check_update
 if ! command -v sudo >/dev/null 2>&1; then
     _yellow "Installing sudo"
